refactor: rename funnies to memes
