+++ src/commands/index.ts
@@ -2,8 +2,8 @@ import { CommandDefinition } from '../lib/command';
 import Logger from '../lib/logger';
 import { ping } from './utils/ping';
 import { whois } from './moderation/whois';
-import { bruheg } from './funnies/bruheg';
-import { boratorium } from './funnies/boratorium';
+import { bruheg } from './memes/bruheg';
+import { boratorium } from './memes/boratorium';
 import { efb } from './a32nx/efb';
 import { deadzones } from './support/deadzones';
 import { screens } from './a32nx/screens';
@@ -26,13 +26,13 @@ import { donate } from './general/donate';
 import { utf8 } from './support/utf-8';
 import { calibrate } from './support/calibrate';
 import { controls } from './support/controls';
-import { nut } from './funnies/nut';
+import { nut } from './memes/nut';
 import { screenshot } from './support/screenshot';
 import { msfs } from './support/msfs';
 import { content } from './support/content';
 import { beginner } from './a32nx/beginner-guide';
 import { briefing } from './a32nx/briefing';
-import { boris } from './funnies/boris';
+import { boris } from './memes/boris';
 import { afloor } from './a32nx/afloor';
 import { airframe } from './a32nx/airframe';
 import { xbox } from './general/xbox';
@@ -49,24 +49,24 @@ import { ctrlE } from './a32nx/ctrlE';
 import { weather } from './a32nx/weather';
 import { metar } from './utils/metar';
 import { qa } from './general/qa';
-import { ptu } from './funnies/ptu';
+import { ptu } from './memes/ptu';
 import { cursor } from './support/cursor';
 import { avatar } from './utils/avatar';
-import { coffee } from './funnies/coffee';
-import { defaultmeme } from './funnies/default';
-import { guard } from './funnies/guard';
-import { juan } from './funnies/juan';
-import { merge } from './funnies/merge';
-import { otter } from './funnies/otter';
-import { p3d } from './funnies/p3d';
-import { poggers } from './funnies/poggers';
-import { pov } from './funnies/pov';
-import { shame } from './funnies/shame';
-import { xp } from './funnies/xp';
+import { coffee } from './memes/coffee';
+import { defaultmeme } from './memes/default';
+import { guard } from './memes/guard';
+import { juan } from './memes/juan';
+import { merge } from './memes/merge';
+import { otter } from './memes/otter';
+import { p3d } from './memes/p3d';
+import { poggers } from './memes/poggers';
+import { pov } from './memes/pov';
+import { shame } from './memes/shame';
+import { xp } from './memes/xp';
 import { station } from './utils/station';
-import { addon } from './funnies/addon';
+import { addon } from './memes/addon';
 import { freetext } from './a32nx/freetext';
-import { cowsay } from './funnies/cowsay';
+import { cowsay } from './memes/cowsay';
 import { synaptic } from './general/synaptic';
 import { directx12 } from './support/directx12';
 import { cfms } from './a32nx/cfms';
@@ -75,23 +75,23 @@ import { discontinuity } from './support/discontinuity';
 import { navdata } from './support/navdata';
 import { build } from './support/build';
 import { tug } from './support/tug';
-import { crak } from './funnies/crak';
+import { crak } from './memes/crak';
 import { weightBalance } from './a32nx/weightBalance';
 import { rules } from './moderation/rules';
 import { fixinfo } from './a32nx/fixinfo';
 import { welcome } from './moderation/welcome';
 import { sop } from './a32nx/sop';
 import { goldenRules } from './general/goldenRules';
-import { fridge } from './funnies/fridge';
+import { fridge } from './memes/fridge';
 import { tiller } from './a32nx/tiller';
 import { assistance } from './a32nx/assistance';
 import { ctd } from './support/ctd';
 import { hud } from './support/hud';
-import { fms } from './funnies/fms';
+import { fms } from './memes/fms';
 import { mcdu } from './a32nx/mcdu';
 import { takeoffPerf } from './a32nx/takeoffPerf';
 import { manualleg } from './support/manualleg';
-import { oim } from './funnies/oim';
+import { oim } from './memes/oim';
 import { wasm } from './support/wasm';
 import { CPDLC } from './a32nx/cpdlc';
 import { simbriefimport } from './a32nx/simbriefimport';
@@ -105,12 +105,12 @@ import { storedWaypoint } from './a32nx/stored-waypoint';
 import { tcas } from './a32nx/tcas';
 import { msfsdisc } from './general/msfsdiscord';
 import { count } from './utils/count';
-import { shomas } from './funnies/shomas';
-import { pw } from './funnies/pw';
+import { shomas } from './memes/shomas';
+import { pw } from './memes/pw';
 import { abbreviations } from './support/abbreviations';
 import { salty } from './general/salty';
 import { airac } from './support/airac';
-import { mico } from './funnies/mico';
+import { mico } from './memes/mico';
 import { translate } from './general/translate';
 import { github } from './a32nx/github';
 import { zulu } from './utils/zulu';
