+++ src/commands/memes/boratorium.ts
@@ -7,7 +7,7 @@ const BORATORIUM_URL = 'https://cdn.discordapp.com/attachments/74757194834247685
 export const boratorium: CommandDefinition = {
     name: 'boratorium',
     description: 'B O R A T',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     executor: async (msg) => {
         const boratoriumEmbed = makeEmbed({ image: { url: BORATORIUM_URL } });
         await msg.channel.send({ embeds: [boratoriumEmbed] });
