+++ src/commands/memes/bruheg.ts
@@ -7,7 +7,7 @@ const BRUHEG_URL = 'https://media.discordapp.net/attachments/772990923486527534/
 export const bruheg: CommandDefinition = {
     name: 'bruheg',
     description: 'bruheg momen',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     executor: async (msg) => {
         const bruhegEmbed = makeEmbed({ image: { url: BRUHEG_URL } });
         await msg.channel.send({ embeds: [bruhegEmbed] });
