+++ src/commands/memes/coffee.ts
@@ -6,7 +6,7 @@ const COFFEE_URL = 'https://www.youtube.com/watch?v=QPfIeVnkZ4Q';
 export const coffee: CommandDefinition = {
     name: 'coffee',
     description: 'Would you like some coffee?',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     executor: async (msg) => {
         await msg.channel.send(COFFEE_URL);
     },
