+++ src/commands/memes/merge.ts
@@ -6,7 +6,7 @@ const MERGE_URL = 'https://tenor.com/view/git-merge-gitmerge-gif-18010083';
 export const merge: CommandDefinition = {
     name: 'merge',
     description: 'GIT MERGE!',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     executor: async (msg) => {
         await msg.channel.send(MERGE_URL);
     },
