+++ src/commands/memes/oim.ts
@@ -5,7 +5,7 @@ const OIM_URL = 'https://cdn.discordapp.com/attachments/879400460891594772/93192
 
 export const oim: CommandDefinition = {
     name: 'oim',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     description: 'oim',
     executor: (msg) => msg.channel.send(OIM_URL),
 };
