+++ src/commands/memes/pov.ts
@@ -7,7 +7,7 @@ const POV_URL = 'https://cdn.discordapp.com/attachments/897491699167793182/99883
 export const pov: CommandDefinition = {
     name: 'pov',
     description: 'Oof',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     executor: async (msg) => {
         const povEmbed = makeEmbed({ image: { url: POV_URL } });
         await msg.channel.send({ embeds: [povEmbed] });
