+++ src/commands/memes/ptu.ts
@@ -7,7 +7,7 @@ const PTU_URL = 'https://cdn.discordapp.com/attachments/898602626436964402/90234
 export const ptu: CommandDefinition = {
     name: 'ptu',
     description: 'Bark',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     executor: async (msg) => {
         const ptuEmbed = makeEmbed({ image: { url: PTU_URL } });
         await msg.channel.send({ embeds: [ptuEmbed] });
