+++ src/commands/memes/pw.ts
@@ -7,7 +7,7 @@ const PW_URL = 'https://cdn.discordapp.com/attachments/748761999747579944/961760
 export const pw: CommandDefinition = {
     name: 'pw',
     description: 'Whale noises',
-    category: CommandCategory.FUNNIES,
+    category: CommandCategory.MEMES,
     executor: async (msg) => {
         const pwEmbed = makeEmbed({ image: { url: PW_URL } });
         await msg.channel.send({ embeds: [pwEmbed] });
